% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuSiE4X.R
\name{SuSiE4X}
\alias{SuSiE4X}
\title{SuSiE4X: Iterative SuSiE fitting for main and interaction effects (dense matrix only)}
\usage{
SuSiE4X(
  X,
  Z = NULL,
  y,
  crossprodX = NULL,
  n_threads = 1,
  Lmain = 5,
  Linteraction = 5,
  max.iter = 15,
  max.eps = 1e-05,
  min.iter = 3,
  susie.iter = 300,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{An n × p matrix of covariates.}

\item{Z}{An n × q matrix of covariates. If Z = NULL, only GxG interaction will be consider.}

\item{y}{An n-vector of responses.}

\item{n_threads}{Integer. Number of threads used in parallel computations, such as blockwise matrix crossproduct or matrix–vector multiplication. Default is 1 (no parallelism). Increasing this can speed up computation on multicore machines when \code{is.large = TRUE}.}

\item{Lmain}{Number of SuSiE effects for main effect X (default = 5).}

\item{Linteraction}{Number of SuSiE effects for interactions (default = 5).}

\item{max.iter}{Maximum number of outer iterations (default = 10).}

\item{max.eps}{Convergence threshold on η (default = 1e-5).}

\item{min.iter}{Minimum number of outer iterations (default = 3).}

\item{susie.iter}{Maximum number of iterations for each SuSiE fit (default = 300).}

\item{verbose}{A logical indicator of whether to display iteration diagnostics (default = TRUE).}

\item{...}{Additional arguments passed to \code{susie_suff_stat}, such as \code{prior_variance}, \code{null_weight}, \code{standardize}, etc. Note that these tuning parameters will only be used in the fine-mapping of marginal effects.}
}
\value{
A list containing results from the iterative SuSiE fitting procedure:
\describe{
  \item{iter}{Total number of outer iterations completed before convergence or reaching maximum.}
  \item{error}{Numeric vector of max change in fitted values across iterations.}
  \item{fitX}{SuSiE result for main effects.}
  \item{fitW}{SuSiE result for interaction terms.}
  \item{fitZ}{Final linear model of y ~ Z with offset.}
  \item{fitJoint}{Final linear model of y ~ Z + credible sets of direct and interaction effects.}
}
}
\description{
This function performs iterative estimation of additive (Z) and main (X) effects, as well as
interaction effects between selected X components and Z, using SuSiE on sufficient statistics.
}
\details{
For large sample size, it uses a blockwise strategy (like in `bam`) to compute `crossprod(X)` efficiently
without requiring the full X'X in memory at once.

The algorithm alternates between updating the effects from Z (via linear projection),
from X (via susie_suff_stat), and interaction terms (via susie_suff_stat after constructing selected interactions).
After convergence, a post-hoc linear model of y ~ Z is fit with offset(etaX + etaW) to estimate
the significance of Z effects adjusted for X and interactions.
}
